LPD-X Merge ClearScrollRequestExecutor/ClearScrollRequestExecutorImpl into pojos
